refactor: move AppInfo to api module; adapt usages
